refactor(library/math): moves "*/operator" to "*/operators"
