+++ src/library/math/operators/constructive/absoluteValueOf/aliases/index.ts
