+++ src/library/math/operators/constructive/absoluteValueOf/aliases/modulusOf.ts
