+++ src/library/math/operators/constructive/absoluteValueOf/definition.declared.ts
