+++ src/library/math/operators/constructive/absoluteValueOf/exports.object.auxiliaries.ts
