+++ src/library/math/operators/constructive/absoluteValueOf/index.ts
