+++ src/library/math/operators/constructive/cofactor.ts
