+++ src/library/math/operators/constructive/greatestCommonDivisor/abbreviations/exports.toolbox.ts
