+++ src/library/math/operators/constructive/greatestCommonDivisor/abbreviations/gcd.ts
