+++ src/library/math/operators/constructive/greatestCommonDivisor/abbreviations/gcf.ts
