+++ src/library/math/operators/constructive/greatestCommonDivisor/abbreviations/index.ts
