+++ src/library/math/operators/constructive/greatestCommonDivisor/aliases/exports.toolbox.ts
