+++ src/library/math/operators/constructive/greatestCommonDivisor/aliases/greatestCommonFactor.ts
