+++ src/library/math/operators/constructive/greatestCommonDivisor/aliases/highestCommonDivisor.ts
