+++ src/library/math/operators/constructive/greatestCommonDivisor/aliases/highestCommonFactor.ts
