+++ src/library/math/operators/constructive/greatestCommonDivisor/aliases/index.ts
