+++ src/library/math/operators/constructive/greatestCommonDivisor/definition.declared.ts
