+++ src/library/math/operators/constructive/greatestCommonDivisor/definition.test.ts
