+++ src/library/math/operators/constructive/greatestCommonDivisor/exports.object.auxiliaries.ts
