+++ src/library/math/operators/constructive/greatestCommonDivisor/exports.object.primary.ts
