+++ src/library/math/operators/constructive/greatestCommonDivisor/index.ts
