+++ src/library/math/operators/constructive/leastCommonMultiple.ts
