+++ src/library/math/operators/hyper/index.ts
