+++ src/library/math/operators/hyper/rank3/index.ts
