+++ src/library/math/operators/hyper/rank3/squared.test.ts
