+++ src/library/math/operators/hyper/rank3/squared.ts
