+++ src/library/math/operators/predicate/index.ts
