+++ src/library/math/operators/predicate/isFinite.test.ts
