+++ src/library/math/operators/predicate/isFinite.ts
