+++ src/library/math/operators/predicate/isNegative.test.ts
