+++ src/library/math/operators/predicate/isNegative.ts
