+++ src/library/math/operators/predicate/isOperable.test.ts
