+++ src/library/math/operators/predicate/isOperable.ts
