+++ src/library/math/operators/predicate/isWhole/aliases/exports.toolbox.ts
