+++ src/library/math/operators/predicate/isWhole/aliases/index.ts
