+++ src/library/math/operators/predicate/isWhole/aliases/isInteger.ts
