+++ src/library/math/operators/predicate/isWhole/definition.declared.ts
