+++ src/library/math/operators/predicate/isWhole/definition.test.ts
