+++ src/library/math/operators/predicate/isWhole/exports.object.auxiliaries.ts
