+++ src/library/math/operators/predicate/isWhole/exports.object.primary.ts
