+++ src/library/math/operators/predicate/isWhole/index.ts
